Some refactoring and simplification of frDonation & frDonor.  These were the first classes in this project and isn't really consistent with how we're writing them now.
Also make it so that if we find a match with an existing contact, we populate the funraise id and save that immediately so that if any mapping errors prevent the record from saving, it still has the funraise id on it for future integration attempts.  It also allows us to remove the update/upsert check that was happening, which led to sometimes hitting duplicate value exceptions in the fr_ID__c field.  Now we will only do upserts on the fr_ID__c.

Also addressed some tech debt where for Campaign supporterId and parentGoalId were primitive long instead of Longs, so we had to set them to 0 instead of null when there wasn't a value.  That's fixed on the platform side so we don't have to check for 0 anymore