Add handling for 2 new properties being sent for a donation: softCreditSupporterId & campaignMappingDisabled
The flag campaignMappingDisabled flag will be set according to the platform configuration, and if it is true then we won't try to match to a Campaign record, but it's still available to be a mapped field.
The softCreditSupporterId will link to the supporter that is the soft credit, displayed in Salesforce as an Opportunity Contact Role (if enabled)
Refactoring error handling.  The error object will now have a picklist displaying the type of object in Funraise and it's record ID, along with the general message field that was there previously
Resolving FUN-7754 & FUN-8533